Fix bad update message
The problem appeared because of bad usage of
_get_tracked_cities_str, which would return either a
comma-separated list of cities OR an 'all cities' string
and was intended to be used for human-readable messages
generation only. The attempt to get a list of cities from
this representation to be used in filtering resulted in
empty update and thus the 'No change' message.
This was mitigated by introducing a _get_tracked_cities
and using it for filtering purposes.

Closes: #4