Fix spamming incident
Global status was not updated in case an exception was raised
during sending a status update message to the users, which resulted
in infinite 'send an update about a change' loop. The exception
was raised because the bot attempted to send a message to the user
who has unsubscribed for good in the first place.
This patch adds proper exception handling (both general exceptions
and the specific one when telegram.error.Unauthorized is raised).
Also some logging to ease further investigations is finally set up.

Closes: #10