React to the Mar 27, 2023 incident
As of midday March 27, 2023 recaptcha with post-verification
exam cities list publication has been introduced. Also there
have been some changes to the page layout, which needed addressing.

This patch introduces a headless browser fetch page method. It is
still not perfect as is hitting recaptcha, but in comparison with
no way to get source page via curl/requests that is a great start.

Related-to: #23