Introduce a fetcher module
Move all webpage html fetching functionality to
the fetcher module. It will be responsible for
pushing html data to the centralized repository as
well.
Checker module will be responsible for html->json
transformation as well as data parsing and filtering.
For now it will rely on html file mounted into the
container to be updated by fetcher module.